Add virtual destructors to LedMatrix and Driver, to prevent memory leaks in case the Driver is deleted by the client code; move configure() to setup() because it calls Hardware to setup pins
