+++ examples/AceSegmentDemo/AceSegmentDemo.ino
@@ -121,6 +121,7 @@ void setup() {
   Serial.begin(115200); // ESP8266 default of 74880 not supported on Linux
   while (!Serial); // Wait until Serial is ready - Leonardo/Micro
 
+  driver->configure();
   renderer
       .setFramesPerSecond(FRAMES_PER_SECOND)
       .configure();
