+++ src/ace_segment/LedMatrixSerial.h
@@ -25,11 +25,16 @@ SOFTWARE.
 #ifndef ACE_SEGMENT_LED_MATRIX_SERIAL_H
 #define ACE_SEGMENT_LED_MATRIX_SERIAL_H
 
-#include "Hardware.h"
 #include "LedMatrix.h"
 
 namespace ace_segment {
 
+class Hardware;
+
+/**
+ * An LedMatrix with an 74HC595 Serial-To-Parallel converter chip on the
+ * segment pins.
+ */
 class LedMatrixSerial: public LedMatrix {
   public:
     LedMatrixSerial(Hardware* hardware, uint8_t numGroups, uint8_t numElements):
@@ -46,42 +51,17 @@ class LedMatrixSerial: public LedMatrix {
       mClockPin = clockPin;
     }
 
-    virtual void configure() override {
-      LedMatrix::configure();
+    virtual void configure() override;
 
-      // TODO: Do I need to set the initial values of the 74HC595?
-      mHardware->pinMode(mLatchPin, OUTPUT);
-      mHardware->pinMode(mDataPin, OUTPUT);
-      mHardware->pinMode(mClockPin, OUTPUT);
+    virtual void enableGroup(uint8_t group) override;
 
-      for (uint8_t group = 0; group < mNumGroups; group++) {
-        uint8_t pin = mGroupPins[group];
-        mHardware->pinMode(pin, OUTPUT);
-        mHardware->digitalWrite(pin, mGroupOff);
-      }
-    }
+    virtual void disableGroup(uint8_t group) override;
 
-    virtual void enableGroup(uint8_t group) override {
-      writeGroupPin(group, mGroupOn);
-    }
-
-    virtual void disableGroup(uint8_t group) override {
-      writeGroupPin(group, mGroupOff);
-    }
-
-    virtual void drawElements(uint8_t pattern) override {
-      mHardware->digitalWrite(mLatchPin, LOW);
-      uint8_t actualPattern = (mElementOn == HIGH) ? pattern : ~pattern;
-      mHardware->shiftOut(mDataPin, mClockPin, MSBFIRST, actualPattern);
-      mHardware->digitalWrite(mLatchPin, HIGH);
-    }
+    virtual void drawElements(uint8_t pattern) override;
 
   protected:
-    /** Write to group pin identified by 'group'. VisibleForTesting. */
-    void writeGroupPin(uint8_t group, uint8_t output) {
-      uint8_t groupPin = mGroupPins[group];
-      mHardware->digitalWrite(groupPin, output);
-    }
+    /** Write to group pin identified by 'group'. */
+    void writeGroupPin(uint8_t group, uint8_t output);
 
     const uint8_t* mGroupPins;
     uint8_t mLatchPin;
